requirements.txt
altair==5.2.0
appscript==1.2.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
docutils==0.20.1
docx2pdf==0.1.8
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
lxml==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
moviepy==1.0.3
nh3==0.2.15
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pkginfo==1.9.6
proglog==0.1.10
protobuf==4.25.1
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pypandoc==1.12
python-dateutil==2.8.2
pytz==2023.3.post1
readme-renderer==42.0
referencing==0.32.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.15.2
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
streamlit-lottie==0.0.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
twine==4.0.2
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wheel==0.42.0
zipp==3.17.0
